feat: implement web worker for stellar scanning
